just use our own variant of unref decay rather than trying to insert into std namespace
